% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cache_dir.R
\name{mr_cache_dir}
\alias{mr_cache_dir}
\title{Check the location of the cache}
\usage{
mr_cache_dir(dir = getOption("mr_cache", default = tools::R_user_dir("mtgr")))
}
\arguments{
\item{dir}{directory to be used as the cache directory}
}
\description{
To change the default location of the cache, use option \code{mr_cache}
(see Examples).
}
\details{
This function is intended to be called internally with no
arguments. It will use the default
location returned by \link[tools:userdir]{tools::R_user_dir} for the application,
\code{mtgr}. This function will also
create the directory if it does not yet exist.
}
\examples{
# Default cache directory
mr_cache_dir()
# Specify a custom cache directory
# (save any existing settings first)
old_cache_setting <- getOption("mr_cache")
options(mr_cache = tempdir())
mr_cache_dir()
# restore existing settings
options(mr_cache = old_cache_setting)
}
