% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_get_17lands_data.R
\name{mr_download_17lands_file}
\alias{mr_download_17lands_file}
\title{Download a 17lands data file}
\usage{
mr_download_17lands_file(
  set,
  data_type,
  event_type,
  use_cache = TRUE,
  overwrite = FALSE,
  cache_dir = mr_cache_dir(),
  quiet = FALSE
)
}
\arguments{
\item{set}{Three-letter set code. Not case-sensitive. For a complete list of
set names and codes, run \code{mr_get_sets()} (not all sets are available on
17lands).}

\item{data_type}{Type of data. Must be one of "game", "draft", or "replay".}

\item{event_type}{Type of event. Must be one of "premier", "traditional",
"sealed", or "quick". Not all formats are available for all event types (in
particular, Quick Draft is only available for one set at the time of
writing).}

\item{use_cache}{Logical; should the cache be used?}

\item{overwrite}{Logical: should any existing file be overwritten?}

\item{cache_dir}{Path to the folder used for caching data.}

\item{quiet}{Logical: should messages emitted by this function be silenced?}
}
\value{
Path to data file
}
\description{
Not all sets or combinations of data types are available. To see the
available files, visit \url{https://www.17lands.com/public_datasets}.
}
\examples{
mr_download_17lands_file("KTK", "game", "premier")
}
