% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_get_17lands.R
\name{mr_get_sets}
\alias{mr_get_sets}
\title{Get 17lands data}
\usage{
mr_get_sets()
}
\value{
Dataframe
}
\description{
Print a dataframe of expansion set codes and names.
}
\details{
Only includes expansion sets that have already been released.
For access to complete set data, see scryr::scry_sets().
}
\examples{
mr_get_sets()
}
