% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{replay_cols}
\alias{replay_cols}
\title{Column names and types in replay data}
\format{
\subsection{\code{replay_cols}}{

A dataframe (tibble) describing columns in 17lands replay data
(\code{data_type = "replay"}).

Columns:
\describe{
\item{regex}{Regular expression to match a column in the replay data}
\item{py_type}{Data type used by python for that column}
\item{arrow}{Data type used by arrow for that column}
...
}
}
}
\source{
\url{https://www.17lands.com/public_datasets}
}
\usage{
replay_cols
}
\description{
Column names and types in replay data
}
\keyword{datasets}
