% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_get_17lands.R
\name{mr_cache_dir}
\alias{mr_cache_dir}
\title{Set the cache directory}
\usage{
mr_cache_dir(dir = tools::R_user_dir("magicr"))
}
\arguments{
\item{dir}{directory to be used as the cache directory}
}
\description{
A configurable default location for persistent data storage.
}
\details{
This function is intended to be called internally with no
arguments. It will use the default
location returned by \link[tools:userdir]{tools::R_user_dir} for the application,
\code{magicr}. This function will also
create the directory if it does not yet exist.
}
\examples{
# Default cache directory
mr_cache_dir()
# Custom cache directory (here, temporary directory)
mr_cache_dir(tempdir())
}
