% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_get_17lands.R
\name{mr_get_17lands_data}
\alias{mr_get_17lands_data}
\title{Get 17lands data}
\usage{
mr_get_17lands_data(
  set,
  data_type,
  event_type,
  nrows = Inf,
  skip = "__auto__",
  select = NULL,
  drop = NULL,
  n_threads = data.table::getDTthreads()
)
}
\arguments{
\item{set}{Set abbreviation. Not case-sensitive. For a list of all set
abbreviations, run mr_get_sets(). Not all sets are available on 17lands.}

\item{data_type}{Type of data. Must be one of "game", "draft", or "replay".}

\item{event_type}{Type of event. Must be one of "premier", "traditional",
or "sealed".}

\item{nrows}{Number of rows to load.}

\item{skip}{Number of rows to skip from top before loading data.}

\item{select}{A vector of column names or numbers to keep; drops the rest.}

\item{drop}{Vector of column names or numbers to drop; keeps the rest.}

\item{n_threads}{The number of threads to use.}
}
\value{
Dataframe
}
\description{
Download data from 17lands and load it into R
}
\details{
Arguments \code{nrows}, \code{skip}, \code{select}, \code{drop}, and \code{n_threads} are passed to
data.table::fread(); for details, see \code{?data.table::fread()}.
}
\examples{
mr_get_17lands_data("KTK", "game", "premier")
}
